cleanup code and renaming files to match naming convention
