+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './styles/index.css';
-import Onboarding from "./onboarding";
+import Onboarding from "./screens/onboarding/Onboarding";
 
 ReactDOM.render(<Onboarding/>, document.getElementById('root'));
