+++ src/styles/Onboarding.css
